支持主题功能（内置深色主题和浅色主题），窗口阴影支持系统阴影，以改善用户体验：
（1）支持主题功能（内置深色主题和浅色主题），并提供主题生成和主题修改的辅助工具
（2）开发使用系统阴影的功能：Windows 10/11和macOS系统默认使用系统阴影，这些系统中系统阴影的体验较好，并且Window下不用使用分层窗口，绘制性能也明显改善。(#415)
（3）支持字体回退功能，从而支持Emoji字体显示
（4）对global.xml文件的结构进行一次重构，支持主题设置，支持字体回退，使得配置文件更加规范，提高可维护性
（5）所有示例程序，均支持主题设置和多语言切换
（6）优化代码的绘制性能，改善淡入淡出动画的绘制性能
（7）修复一些缺陷，尤其是控件在滚动条区域内的兼容性缺陷。

+++ duilib/Core/UiRect.h
@@ -53,11 +53,11 @@ public:
 
     /** 获取中心点的X轴坐标值
     */
-    int32_t CenterX() const { return (left + right) / 2; }
+    int32_t CenterX() const { return (int32_t)(((int64_t)left + (int64_t)right) / 2); }
 
     /** 获取中心点的Y轴坐标值
     */
-    int32_t CenterY() const { return (top + bottom) / 2; }
+    int32_t CenterY() const { return (int32_t)(((int64_t)top + (int64_t)bottom) / 2); }
 
     /** 获取矩形中心点坐标（X轴坐标和Y轴坐标）
     */
